--- DNDCore/src/main/java/com/ericarao/dnd/core/model/AbstractPlayerLogin.java
@@ -1,16 +1,0 @@
-package com.ericarao.dnd.core.model;
-
-import org.immutables.value.Value;
-
-@DndStyle
-@Value.Immutable
-public abstract class AbstractPlayerLogin implements NetworkPacket {
-    public abstract String getRoomName();
-    public abstract String getRoomPassword();
-    public abstract String getPlayerName();
-
-    @Override
-    public PacketType getType() {
-        return PacketType.PlayerLogin;
-    }
-}
