--- DNDCore/src/main/java/com/ericarao/dnd/core/model/AbstractPlayerLoginCredentials.java
@@ -1,12 +1,0 @@
-package com.ericarao.dnd.core.model;
-
-import org.immutables.value.Value;
-
-@DndStyle
-@Value.Immutable
-public abstract class AbstractPlayerLoginCredentials{
-    public abstract String roomName();
-    public abstract String dmIP();
-    public abstract String playerName();
-    public abstract String roomPassword();
-}
