--- DNDCore/src/main/java/com/ericarao/dnd/core/model/AbstractPlayerLoginResponse.java
@@ -1,14 +1,0 @@
-package com.ericarao.dnd.core.model;
-
-import org.immutables.value.Value;
-
-@DndStyle
-@Value.Immutable
-public abstract class AbstractPlayerLoginResponse implements NetworkPacket {
-    public abstract boolean getSuccess();
-
-    @Override
-    public PacketType getType() {
-        return PacketType.PlayerLoginResponse;
-    }
-}
