--- DNDCore/src/main/java/com/ericarao/dnd/core/model/PacketType.java
@@ -1,10 +1,0 @@
-package com.ericarao.dnd.core.model;
-
-public enum PacketType {
-    RegisterPlayer,
-    ClientUpdate,
-    PlayerLogin,
-    PlayerLoginResponse,
-    PlayerUpdateStatsDM,
-    RegisterPlayerResponse
-}
