--- DNDCore/src/main/java/com/ericarao/dnd/core/model/ServerResponseIF.java
@@ -1,12 +1,0 @@
-package com.ericarao.dnd.core.model;
-
-import org.immutables.value.Value;
-
-import java.util.Set;
-
-@DndStyle
-@Value.Immutable
-public interface ServerResponseIF {
-    Set<Integer> getClientIds();
-    NetworkPacket getResponse();
-}
