+++ src/main/java/com/ericarao/dnd/View/LoginStatusView.java
@@ -1,0 +1,36 @@
+package com.ericarao.dnd.View;
+
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
+import javafx.scene.Scene;
+import javafx.scene.layout.GridPane;
+import javafx.scene.text.Font;
+import javafx.scene.text.FontWeight;
+import javafx.scene.text.Text;
+
+public class LoginStatusView {
+
+    private Scene scene;
+
+    public Scene getScene() {
+        if (scene == null) {
+            scene = initScene();
+        }
+        return scene;
+    }
+
+    private Scene initScene() {
+        GridPane grid = new GridPane();
+        grid.setAlignment(Pos.CENTER);
+        grid.setHgap(10);
+        grid.setVgap(10);
+        grid.setPadding(new Insets(25, 25, 25, 25));
+
+        Text scenetitle = new Text("Trying to Login");
+
+        scenetitle.setFont(Font.font("Tahoma", FontWeight.NORMAL, 20));
+        grid.add(scenetitle, 0, 0, 2, 1);
+
+        return new Scene(grid, 500, 500);
+    }
+}
