+++ src/main/java/com/ericarao/dnd/model/AbstractClientUpdate.java
@@ -1,0 +1,18 @@
+package com.ericarao.dnd.model;
+
+import org.immutables.value.Value;
+
+@DndStyle
+@Value.Immutable
+public abstract class AbstractClientUpdate implements NetworkPacket {
+    public abstract int getDamage();
+    public abstract String getStatusEffect();
+    public abstract int getSavingThrow();
+    public abstract int getCurrentHP();
+
+    @Override
+    @Value.Derived
+    public PacketType getType() {
+        return PacketType.ClientUpdate;
+    }
+}
