+++ src/main/java/com/ericarao/dnd/model/AbstractPlayerLoginCredentials.java
@@ -1,0 +1,12 @@
+package com.ericarao.dnd.model;
+
+import org.immutables.value.Value;
+
+@DndStyle
+@Value.Immutable
+public abstract class AbstractPlayerLoginCredentials{
+    public abstract String roomName();
+    public abstract String dmIP();
+    public abstract String playerName();
+    public abstract String roomPassword();
+}
