+++ src/main/java/com/ericarao/dnd/model/AbstractPlayerLoginResponse.java
@@ -1,0 +1,14 @@
+package com.ericarao.dnd.model;
+
+import org.immutables.value.Value;
+
+@DndStyle
+@Value.Immutable
+public abstract class AbstractPlayerLoginResponse implements NetworkPacket {
+    public abstract boolean getSuccess();
+
+    @Override
+    public PacketType getType() {
+        return PacketType.PlayerLoginResponse;
+    }
+}
