+++ src/main/java/com/ericarao/dnd/model/PacketType.java
@@ -1,0 +1,10 @@
+package com.ericarao.dnd.model;
+
+public enum PacketType {
+    RegisterPlayer,
+    ClientUpdate,
+    PlayerLogin,
+    PlayerLoginResponse,
+    PlayerUpdateStatsDM,
+    RegisterPlayerResponse
+}
