+++ src/main/java/com/ericarao/dnd/model/ServerResponseIF.java
@@ -1,0 +1,12 @@
+package com.ericarao.dnd.model;
+
+import org.immutables.value.Value;
+
+import java.util.Set;
+
+@DndStyle
+@Value.Immutable
+public interface ServerResponseIF {
+    Set<Integer> getClientIds();
+    NetworkPacket getResponse();
+}
