+++ src/main/java/com/ericarao/dnd/utils/JsonUtils.java
@@ -1,0 +1,25 @@
+package com.ericarao.dnd.utils;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+
+public class JsonUtils {
+    public static final ObjectMapper MAPPER = buildObjectMapper();
+
+    private static ObjectMapper buildObjectMapper() {
+        // Jackson's Object Mapper handles serializing objects to JSON
+        // and deserializing JSON to their corresponding Java objects.
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new Jdk8Module());
+        mapper.registerModule(new JavaTimeModule());
+        mapper.configure(JsonParser.Feature.ALLOW_COMMENTS, false);
+        mapper.configure(JsonParser.Feature.AUTO_CLOSE_SOURCE, true);
+        mapper.configure(JsonGenerator.Feature.AUTO_CLOSE_TARGET, true);
+        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        return mapper;
+    }
+}
